Deleted unused html-files

+++ src/test/java/com/example/bilabonnement/repository/EconomyRepositoryTest.java
@@ -7,7 +7,8 @@ public class EconomyRepositoryTest extends TestCase {
     public void testTotalMonthlyIncomeThisMonthFromRentedCars() {
         //Arrange
         EconomyRepository economyRepo = new EconomyRepository();
-        int expectedResult = 30;
+        //Vi er klar over at der ikke kommer tallet 30 ud, da de vil ændre sig alt efter hvor mange og hvilke lejeaftaler der er oprettet
+        double expectedResult = 30;
         //Act
         double actualResult = economyRepo.totalMonthlyIncomeThisMonthFromRentedCars();
